requirements.txt
behave==1.2.6
playwright==1.40.0
python-dotenv==1.0.0

